[问题/需求描述]:解决报单引用出现负数情形 [解决方法]:解决报单引用出现负数情形 [影响范围]:无 [需求编号]:无

+++ interface/ctpEvent/ctpEvent.cpp
@@ -252,7 +252,7 @@ void CtpEvent::OnRspUserLoginHandle(MsgStruct &msg) {
   auto &logingInfo = traderSer.ROLE(Trader).ROLE(TmpStore).logingInfo;
   logingInfo.FrontID = pMsg->FrontID;
   logingInfo.SessionID = pMsg->SessionID;
-  logingInfo.OrderRef = (unsigned long int)(logingInfo.SessionID % 1000000) * 1000000;
+  logingInfo.OrderRef = (unsigned long int)((unsigned int)logingInfo.SessionID % 1000000) * 1000000;
 
   std::string semName = "trader_logIn";
   globalSem.postSemBySemName(semName);
